Set perf to 4M max size and running for max 500msecs

+++ pre-post-scripts/stop-perf.sh
@@ -2,20 +2,13 @@
 
 LOG_DIR=$(jq -r .log_dir /tmp/interop-variables.json)
 
-pkill -SIGINT -f 'perf record'
-
-sleep 3 # wait for perf to finish
-
-if [ -d ${LOG_DIR} ]; then
-    echo "perf data directory exists"
+if pgrep -f 'perf record' > /dev/null; then
+    echo "A perf record process is already running. Stopping it..."
+    pkill -SIGINT -f 'perf record'
+    sleep 1
 else
-    echo "perf data directory does not exist"
-    exit 127
+    echo "No existing perf record process found."
 fi
 
-touch ${LOG_DIR}/perf.json
-
-echo "perf data succeeded"
-sudo perf script -i ${LOG_DIR}/perf.data > ${LOG_DIR}/perf_record.txt
-sudo rm ${LOG_DIR}/perf.data
-
+sudo perf script -i "${LOG_DIR}/perf.data" > "${LOG_DIR}/perf_record.txt"
+sudo rm "${LOG_DIR}/perf.data"
